feat: Implement §14a EnWG curtailment constraints for charging points and heat pumps.
- Added support for §14a EnWG curtailment settings in the to_powermodels function.
- Introduced new variables and constraints in the OPF to handle curtailment logic.
- Enhanced logging for curtailment statistics in the from_powermodels function.
- Updated the objective function to minimize costs associated with §14a curtailment.
- Ensured all time-varying data in the PyPSA network is of float64 dtype to prevent errors.
- Refactored various functions to improve readability and maintainability.
- Added new Julia files for §14a constraints and integrated them into the OPF workflow.

+++ edisgo/io/electromobility_import.py
@@ -846,7 +846,8 @@ def distribute_private_charging_demand(edisgo_obj):
     designated_charging_point_capacity_df = pd.DataFrame(
         index=user_centric_weights_df.index,
         columns=["designated_charging_point_capacity"],
-        data=0,
+        data=0.0,
+        dtype=float,
     )
 
     for destination in private_charging_df.destination.sort_values().unique():
@@ -986,7 +987,8 @@ def distribute_public_charging_demand(edisgo_obj, **kwargs):
     designated_charging_point_capacity_df = pd.DataFrame(
         index=grid_and_user_centric_weights_df.index,
         columns=["designated_charging_point_capacity"],
-        data=0,
+        data=0.0,
+        dtype=float,
     )
 
     columns = [
